feat(reports): move export functions into a separate file
